Folgende Anforderungen seitens des Gym Neufeld wurden umgesetzt:

- Neuer Berechnungsmodus gemäss Excel des Kantons
- In der Pensenmeldung soll die IPB-Korrektur in einer eigenen Zeile ausgewiesen werden und nicht mehr einer Teilanstellung hinzugerechnet werden
- Die in SAP einzutragenden Daten sollen sowohl im Pensenblatt als auch im CSV-Export direkt sichtbar sein (sodass das Excel nicht mehr verwendet werden muss)

+++ install-neufeld.sh
@@ -1,7 +1,15 @@
 #!/bin/sh
-yarn build
-SERVER=pm.gymneufeld.site
-ROOT=/var/www/pensen.gymneufeld.site
+
+#Build ist am Gym Neufeld vom Deployment getrennt
+#yarn build
+
+if [ -z "$1" ] || [ -z "$2" ]; then
+  exit 1
+fi
+
+SERVER="$1"
+ROOT="$2"
+
 ME=$(ssh ${SERVER} echo \$\{USER\})
 ssh ${SERVER} sudo chown -R ${ME} ${ROOT}
 rsync -av --delete dist/* ${SERVER}:${ROOT}
